Resolução de quase todos os exercícios do Capítulo 7.

+++ ch07/planner/routes/events.py
@@ -2,6 +2,7 @@ from beanie import PydanticObjectId
 from fastapi import APIRouter, Depends, HTTPException, Request,status
 from models.events import Event, EvenUpdate
 from database.connection import Database
+from auth.authenticate import authenticate
 
 event_router = APIRouter(tags=["Events"])
 
@@ -20,21 +21,28 @@ async def retrieve_event(id: PydanticObjectId) -> Event:
     return event
 
 @event_router.post("/new")
-async def create_event(body: Event) -> dict:
+async def create_event(body: Event, user: str = Depends(authenticate)) -> dict:
+    body.creator = user
     await event_database.save(body)
     return {
         "message": "Event created successfully."
     }
 
 @event_router.delete("/delete/{id}")
-async def delete_event(id: PydanticObjectId) -> dict:
+async def delete_event(id: PydanticObjectId, user: str = Depends(authenticate)) -> dict:
+    event = await event_database.get(id)
+    if event.creator != user:
+        raise HTTPException(status_code=status.HTTP_404_NOT_FOUND, detail="Event not found.")
     event = await event_database.delete(id)
     if not event:
         raise HTTPException(status_code=status.HTTP_404_NOT_FOUND, detail="Event with supplied ID does not exist.")
     return {"message": "Event deleted successfully."}
 
 @event_router.put("/{id}", response_model=Event)
-async def update_event(id: PydanticObjectId, body: EvenUpdate) -> Event:
+async def update_event(id: PydanticObjectId, body: EvenUpdate, user: str = Depends(authenticate)) -> Event:
+    event = await event_database.get(id)
+    if event.creator != user:
+        raise HTTPException(status_code=status.HTTP_400_BAD_REQUEST, detail="Operation not allowed.")
     update_event = await event_database.update(id, body)
     if not update_event:
         raise HTTPException(status_code=status.HTTP_404_NOT_FOUND, detail="Event with supplied ID does not exist.")
